more hacking on depots and vehicles
added and then commented out statements that printed out the state of
wait and transit times.  those transit times are indeed long.

tweaked the code for whether or not vehicles return to the same
depot.  Just randomizing the list breaks things.  Didn't explore
that.  Instead, if vehicles end up at a different depot than start,
then I split the depots in half, round robin both the starts and ends
to the vehicles.  One way trips.

upped the size of the problem too.