hacking
revert to tours, fix code of a typo

tried and aborted removing the push back of 6 hours from the latest
time variable for starting pickup windows.  Not sure how to deal with
this better.

trying to fix the occasional crash due to time extending beyond 24hr
horizon by sticking in a "hard_stop" variable.

Realizing that one of the big slowdowns for the code is computing the
all to all distance matrix.  n**2 operation, so with goes up by a
factor of 10, that operation goes up by a factor of 1000

also dropped back to 1000 vehicle problem size